requirements.txt
asgiref==3.3.1
Django==3.1.5
django-ckeditor==5.4.0
django-crispy-forms==1.10.0
django-js-asset==1.0.0
django-taggit==1.2.0
Pillow==8.1.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
